+++ gdal_subsetter/__main__.py
@@ -5,33 +5,31 @@ __main__.py
 
 Runs the harmony_gdal CLI
 """
-import argparse
-import logging
-import time
 import os
+from argparse import ArgumentParser
 
-import harmony
+from harmony import is_harmony_cli, run_cli, setup_cli
+
 from .transform import HarmonyAdapter
 
 
 def main():
-    """
-    Parses command line arguments and invokes the appropriate method to respond to them
+    """Parses command line arguments and invokes the appropriate method to respond to them
 
     Returns
     -------
     None
     """
 
-    parser = argparse.ArgumentParser(
-        prog='harmony-gdal', description='Run the GDAL service'
+    parser = ArgumentParser(
+        prog='harmony-gdal-adapter', description='Run the GDAL service',
     )
 
-    harmony.setup_cli(parser)
-    args = parser.parse_args()
+    setup_cli(parser)
+    harmony_args = parser.parse_args()
 
-    if (harmony.is_harmony_cli(args)):
-        harmony.run_cli(parser, args, HarmonyAdapter)
+    if (is_harmony_cli(harmony_args)):
+        run_cli(parser, harmony_args, HarmonyAdapter)
     else:
         parser.error("Only --harmony CLIs are supported")